Protect notes with middleware

+++ backend/src/controllers/notes.ts
@@ -1,10 +1,15 @@
 import { RequestHandler } from 'express';
 import NoteModel from '../models/note';
 import createHttpError from 'http-errors';
+import { assertIsDefined } from '../utils/assertIsDefined';
 
 export const getNotes: RequestHandler = async (req, res) => {
+  const authUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authUserId);
+
+    const notes = await NoteModel.find({ userId: authUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     console.log(error);
@@ -14,13 +19,17 @@ export const getNotes: RequestHandler = async (req, res) => {
 
 export const getNote: RequestHandler = async (req, res) => {
   const noteId = req.params.noteId;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefined(authUserId);
+
     const note = await NoteModel.findById(noteId).exec();
 
-    if (!note) {
-      throw createHttpError(404, 'Note Not Found!');
-    }
+    if (!note) throw createHttpError(404, 'Note Not Found!');
+
+    if (!note.userId.equals(authUserId))
+      throw createHttpError(401, 'You can not access this note');
 
     res.status(200).json(note);
   } catch (error) {
@@ -42,11 +51,15 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefined(authUserId);
+
     if (!title) throw createHttpError(400, 'Note is missing a title!');
 
     const newNote = await NoteModel.create({
+      userId: authUserId,
       title: title,
       text: text,
     });
@@ -76,13 +89,19 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefined(authUserId);
+
     if (!newTitle) throw createHttpError(400, 'Note missing a title!');
 
     const note = await NoteModel.findById(noteId).exec();
 
     if (!note) throw createHttpError(404, 'Note Not Found!');
+
+    if (!note.userId.equals(authUserId))
+      throw createHttpError(401, 'You can not update this note');
 
     note.title = newTitle;
     note.text = newText;
@@ -98,10 +117,16 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefined(authUserId);
+
     const note = await NoteModel.findById(noteId).exec();
     if (!note) throw createHttpError(404, 'Note Not Found!');
+
+    if (!note.userId.equals(authUserId))
+      throw createHttpError(401, 'You can not delete this note');
 
     await note.deleteOne();
     res.sendStatus(204);
